feat: split accounts api and services
